* uploading changed attachments * fixed bug with lastModifiedVersion if no changes occured

+++ ZoteroSentience/src/sk/mung/sentience/zoterosentience/storage/ZoteroSync.java
@@ -2,26 +2,34 @@ package sk.mung.sentience.zoterosentience.storage;
 
 import org.xmlpull.v1.XmlPullParserException;
 
+import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import sk.mung.sentience.zoteroapi.UploadStatus;
 import sk.mung.sentience.zoteroapi.Zotero;
 import sk.mung.sentience.zoteroapi.entities.CollectionEntity;
+import sk.mung.sentience.zoteroapi.entities.Field;
+import sk.mung.sentience.zoteroapi.entities.Item;
 import sk.mung.sentience.zoteroapi.entities.ItemEntity;
+import sk.mung.sentience.zoteroapi.entities.ItemField;
 
 public class ZoteroSync
 {
-    private ZoteroStorage storage;
-    private Zotero zotero;
+    private final ZoteroStorage storage;
+    private final Zotero zotero;
+    private final File downloadDir;
     
-    public ZoteroSync(ZoteroStorage storage, Zotero zotero)
+    public ZoteroSync(ZoteroStorage storage, Zotero zotero, File downloadDir)
     {
         this.storage = storage;
         this.zotero = zotero;
+        this.downloadDir = downloadDir;
     }
     
-    public void syncCollections() throws IOException, XmlPullParserException
+    private void syncCollections() throws IOException, XmlPullParserException
     {
         int version = storage.getCollectionsVersion();
         Map<String, Integer> collectionVersions = zotero.getCollectionsVersions(version);
@@ -36,7 +44,7 @@ public class ZoteroSync
         storage.setCollectionsVersion(lastModifiedVersion);
     }
     
-    public void syncDeletions() throws IOException, XmlPullParserException
+    private void syncDeletions() throws IOException, XmlPullParserException
     {
         int version = storage.getDeletionsVersion();
         Map<String, List<String>> deletions = zotero.getDeletions(version);
@@ -45,21 +53,34 @@ public class ZoteroSync
         if( deletions.containsKey("collections") )
         {
         	List<String> collections = deletions.get("collections");
-              
-            storage.deleteCollections( collections);
+            if(collections.size() > 0)
+            {
+                storage.deleteCollections( collections);
+            }
         }
 
         if( deletions.containsKey("items") )
         {
             List<String> items = deletions.get("items");
+            if(items.size() > 0)
+            {
+                storage.deleteItems( items);
+            }
+        }
 
-            storage.deleteItems( items);
+        if( deletions.containsKey("tags") )
+        {
+            List<String> tags = deletions.get("tags");
+            if(tags.size() > 0)
+            {
+                storage.deleteTags(tags);
+            }
         }
         
         storage.setDeletionsVersion(lastModifiedVersion);
     }
 
-    public void syncItems() throws IOException, XmlPullParserException
+    private void syncItems() throws IOException, XmlPullParserException
     {
         int version = storage.getItemsVersion();
         Map<String, Integer> itemVersions = zotero.getItemsVersions(version);
@@ -74,5 +95,94 @@ public class ZoteroSync
         storage.setItemsVersion(lastModifiedVersion);
     }
 
+    public void fullSync() throws IOException, XmlPullParserException
+    {
+        int collectionsVersion;
+        int deletionsVersion;
+        do
+        {
+            syncCollections();
+            syncItems();
+            syncDeletions();
+            collectionsVersion = storage.getCollectionsVersion();
+            deletionsVersion = storage.getDeletionsVersion();
+        } while( collectionsVersion != deletionsVersion);
+        uploadAttachments();
+    }
 
+    private void uploadAttachments()
+    {
+        List<ItemFilePair> pairs =  scanForUpdates();
+        for( ItemFilePair pair : pairs)
+        {
+            UploadStatus status = zotero.uploadAttachment(pair.getFile(), pair.getItem());
+            if(status == UploadStatus.SUCCESS)
+            {
+                pair.getItem()
+                        .getField(ItemField.MODIFICATION_TIME)
+                        .setValue(Long.toString(pair.getFile().lastModified()));
+            }
+            else
+            {
+                //TODO: resolve conflicts or out of storage
+            }
+        }
+    }
+
+    private class ItemFilePair
+    {
+        private final File file;
+        private final Item item;
+
+
+        private ItemFilePair(File file, Item item)
+        {
+            this.file = file;
+            this.item = item;
+        }
+
+        private File getFile()
+        {
+            return file;
+        }
+
+        private Item getItem()
+        {
+            return item;
+        }
+    }
+    private List<ItemFilePair> scanForUpdates()
+    {
+        List<ItemFilePair> pairs = new ArrayList<ItemFilePair>();
+        if(downloadDir.exists())
+        {
+            File[] keys = downloadDir.listFiles();
+            if(null!=keys)
+            {
+                for(File keyDir : keys)
+                {
+                    File[] attachments = keyDir.listFiles();
+                    if(null != attachments && attachments.length > 0)
+                    {
+                        long modificationTime = attachments[0].lastModified();
+                        Item item = storage.findItemByKey(keyDir.getName());
+                        if(item != null)
+                        {
+                            Field field = item.getField(ItemField.MODIFICATION_TIME);
+                            if(field != null)
+                            {
+                                long libraryTime = Long.valueOf(field.getValue());
+                                if(modificationTime > libraryTime)
+                                {
+                                    pairs.add(new ItemFilePair(attachments[0],item));
+                                }
+                            }
+                        }
+                    }
+
+                }
+            }
+        }
+        return pairs;
+    }
 }
